[feat] #6 ConnectedManagable 프로토콜 채택
